refactor: Move condition string representation to the datetime.h/cc
