+++ src/baldr/datetime.cc
@@ -702,6 +702,100 @@ from_cache(const TP& tp,
   infos.emplace_back(tp.get_info());
   return infos.back();
 }
+// Format dow mask into human-readable format. Week starts from Sunday in the mask
+// - 0b1000001 => Su,Sa
+// - 0b0111110 => Mo-Fr
+// - 0b0101110 => Mo-We,Fr
+void format_dow(uint64_t dow_mask, std::ostringstream& ss) {
+  const std::string days[] = {"Su", "Mo", "Tu", "We", "Th", "Fr", "Sa"};
+
+  // state machine that helps us to understand the range
+  bool prev = false;
+  bool preprev = false;
+  // flag to track do we need a comma or not
+  bool empty = true;
+
+  for (int i = 0; i < 7; ++i) {
+    const bool curr = dow_mask & (1 << i);
+
+    if (curr && !prev) {
+      // range start
+      if (!empty) {
+        ss << ',';
+      }
+      ss << days[i];
+      empty = false;
+    }
+
+    if (!curr && prev && preprev) {
+      // we passed the range end - print the previous day
+      ss << '-' << days[i - 1];
+    }
+
+    preprev = prev;
+    prev = curr;
+  }
+
+  // handle range ending after loop
+  if (prev && preprev) {
+    // we passed the range end - print the previous day
+    ss << '-' << days[6];
+  }
+}
+
+std::string month_name(valhalla::baldr::MONTH month) {
+  using valhalla::baldr::MONTH;
+  switch (month) {
+    case MONTH::kJan:
+      return "Jan";
+    case MONTH::kFeb:
+      return "Feb";
+    case MONTH::kMar:
+      return "Mar";
+    case MONTH::kApr:
+      return "Apr";
+    case MONTH::kMay:
+      return "May";
+    case MONTH::kJun:
+      return "Jun";
+    case MONTH::kJul:
+      return "Jul";
+    case MONTH::kAug:
+      return "Aug";
+    case MONTH::kSep:
+      return "Sep";
+    case MONTH::kOct:
+      return "Oct";
+    case MONTH::kNov:
+      return "Nov";
+    case MONTH::kDec:
+      return "Dec";
+    default:
+      return "";
+  }
+}
+
+std::string dow_name(valhalla::baldr::DOW dow) {
+  using valhalla::baldr::DOW;
+  switch (dow) {
+    case DOW::kSunday:
+      return "Su";
+    case DOW::kMonday:
+      return "Mo";
+    case DOW::kTuesday:
+      return "Tu";
+    case DOW::kWednesday:
+      return "We";
+    case DOW::kThursday:
+      return "Th";
+    case DOW::kFriday:
+      return "Fr";
+    case DOW::kSaturday:
+      return "Sa";
+    default:
+      return "";
+  }
+}
 } // namespace
 
 using namespace valhalla::baldr;
@@ -1172,6 +1266,85 @@ bool is_conditional_active(const bool type,
     dt_in_range = (dt_in_range && time_in_range);
   } catch (std::exception& e) {}
   return (dow_in_range && dt_in_range);
+}
+
+std::string conditional_to_str(const bool type,
+                               const uint8_t begin_hrs,
+                               const uint8_t begin_mins,
+                               const uint8_t end_hrs,
+                               const uint8_t end_mins,
+                               const uint8_t dow,
+                               const uint8_t begin_week,
+                               const uint8_t begin_month,
+                               const uint8_t begin_day_dow,
+                               const uint8_t end_week,
+                               const uint8_t end_month,
+                               const uint8_t end_day_dow) {
+  std::ostringstream ss;
+
+  // Handle begin_day_dow and end_day_dow type, day of the month or Nth weekday
+  bool need_space = false;
+  if (type == kYMD) { // kYMD - use days of the month
+    if (begin_month != 0) {
+      ss << month_name(static_cast<MONTH>(begin_month));
+      if (begin_day_dow != 0) {
+        ss << ' ' << static_cast<int>(begin_day_dow);
+      }
+      need_space = true;
+    }
+    if (end_month != 0 && (end_month != begin_month || end_day_dow != begin_day_dow)) {
+      if (end_month != begin_month || end_day_dow != 0) {
+        ss << '-';
+      }
+      if (end_month != 0) {
+        ss << month_name(static_cast<MONTH>(end_month));
+        if (end_day_dow != 0) {
+          ss << ' ' << static_cast<int>(end_day_dow);
+        }
+        need_space = true;
+      }
+    }
+  } else { // kNthDow - use Nth weekday
+    if (begin_month != 0) {
+      ss << month_name(static_cast<MONTH>(begin_month)) << ' ';
+    }
+    if (begin_week != 0) {
+      ss << static_cast<int>(begin_week) << ' ';
+    }
+    ss << dow_name(static_cast<DOW>(begin_day_dow));
+    if (end_day_dow != 0 &&
+        (end_day_dow != begin_day_dow || end_week != begin_week || end_month != begin_month)) {
+      ss << '-';
+      if ((end_week != 0)) {
+        ss << static_cast<int>(end_week) << ' ';
+      }
+      ss << dow_name(static_cast<DOW>(end_day_dow));
+    }
+    need_space = true;
+  }
+
+  if (dow) {
+    if (need_space) {
+      ss << ' ';
+    }
+    format_dow(dow, ss);
+    need_space = true;
+  }
+
+  // Handle time range
+  if (begin_hrs != 0 || end_hrs != 0 || begin_mins != 0 || end_mins != 0) {
+    if (need_space) {
+      ss << ' ';
+    }
+    ss << std::setw(2) << std::setfill('0') << static_cast<int>(begin_hrs) << ':' << std::setw(2)
+       << std::setfill('0') << static_cast<int>(begin_mins);
+    if (end_hrs != 0 || end_mins != 0) {
+      ss << '-' << std::setw(2) << std::setfill('0') << static_cast<int>(end_hrs) << ':'
+         << std::setw(2) << std::setfill('0') << static_cast<int>(end_mins);
+    }
+  }
+
+  return ss.str();
 }
 
 uint32_t second_of_week(uint32_t epoch_time, const date::time_zone* time_zone) {
